requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
blinker==1.9.0
blis==1.2.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
confection==0.1.5
cryptography==44.0.1
cymem==2.0.11
fastapi==0.115.8
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langcodes==3.5.0
language_data==1.3.0
lxml==5.3.1
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.12
narwhals==1.27.1
nltk==3.9.1
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pdfminer.six==20240706
pillow==11.1.0
preshed==3.0.9
protobuf==5.29.3
pyarrow==19.0.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.23.1
scikit-learn==1.6.1
scipy==1.15.2
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.45.3
streamlit==1.42.2
tenacity==9.0.0
thinc==8.3.4
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.2
